requirements.txt
Django==2.2.4
gunicorn==19.4.5
psycopg2==2.7.1
dj-database-url==0.4.0
whitenoise==4.1.4
djangorestframework==3.10.2
